requirements.txt
asgiref==3.7.2
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
importlib-metadata==6.8.0
inflection==0.5.1
Markdown==3.4.4
packaging==23.1
psycopg2-binary==2.9.7
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing-extensions==4.7.1
uritemplate==4.1.1
zipp==3.16.2
